dom/html/test/test_bug383383.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=383383
-->
<head>
  <title>Test for Bug 383383</title>
  <script type="text/javascript" src="/MochiKit/MochiKit.js"></script>
  <script src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=383383">Mozilla Bug 383383</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript" for=" window " event=" onload() ">

var foo = "bar";

</script>

<script class="testbody" type="text/javascript" for="object" event="handler">

// This script should fail to run
foo = "baz";

isnot(foo, "baz", "test failed");

</script>

<script class="testbody" type="text/javascript">

ok(foo == "bar", "test passed");

</script>
</pre>
</body>
</html>
